+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `declare_slash_commands` to the Client abstract interface.
 - Client and Component are now bound to a specific event loop with said loop being exposed by a `loop` property.
 - `BaseSlashCommand.tracked_command`.
+- Upgraded `load_modules`, `unload_modules` and `reload_modules` to the Client abstract interface.
+- `Component.make_loader` shorthand method for making a module loader and unloader for a component.
+- `tanjun.abc.AbstractLoader` to make loaders more standard and easier to custom implement.
 
 ### Changed
 - Renamed `Client.clear_commands` to `Client.clear_slash_commands`.